[ANDROSDK-444] Move all to maintenance package
