+++ core/src/main/java/org/hisp/dhis/android/core/maintenance/ForeignKeyViolation.java
@@ -26,7 +26,7 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package org.hisp.dhis.android.core.systeminfo;
+package org.hisp.dhis.android.core.maintenance;
 
 import android.database.Cursor;
 import android.support.annotation.NonNull;
